server/requirements.txt
aiocache==0.12.3
fastapi==0.115.11
httpx==0.28.1
python-dotenv==1.0.1
uvicorn==0.34.0